+++ scripts/ipadapter.py
@@ -1,3 +1,4 @@
+import json
 from PIL import Image
 import gradio as gr
 from modules import scripts, processing, shared, ipadapter
@@ -39,6 +40,9 @@ class Script(scripts.Script):
         num_units = num_units or 1
         return (num_units * [gr.update(visible=True)]) + ((MAX_ADAPTERS - num_units) * [gr.update(visible=False)])
 
+    def display_advanced(self, advanced):
+        return [gr.update(visible=advanced), gr.update(visible=advanced)]
+
     def ui(self, _is_img2img):
         with gr.Accordion('IP Adapters', open=False, elem_id='ipadapter'):
             units = []
@@ -72,7 +76,11 @@ class Script(scripts.Script):
                     masks[i].change(fn=self.load_images, inputs=[masks[i]], outputs=[mask_galleries[i]])
                 units.append(unit)
             num_adapters.change(fn=self.display_units, inputs=[num_adapters], outputs=units)
-        return [num_adapters] + adapters + scales + files + starts + ends + masks
+            layers_active = gr.Checkbox(label='Layer options', default=False, interactive=True)
+            layers_label = gr.HTML('<a href="https://huggingface.co/docs/diffusers/main/en/using-diffusers/ip_adapter#style--layout-control" target="_blank">InstantStyle: advanced layer activation</a>', visible=False)
+            layers = gr.Text(label='Layer scales', placeholder='{\n"down": {"block_2": [0.0, 1.0]},\n"up": {"block_0": [0.0, 1.0, 0.0]}\n}', rows=1, type='text', interactive=True, lines=5, visible=False, show_label=False)
+            layers_active.change(fn=self.display_advanced, inputs=[layers_active], outputs=[layers_label, layers])
+        return [num_adapters] + adapters + scales + files + starts + ends + masks + [layers_active] + [layers]
 
     def process(self, p: processing.StableDiffusionProcessing, *args): # pylint: disable=arguments-differ
         if shared.backend != shared.Backend.DIFFUSERS:
@@ -94,4 +102,11 @@ class Script(scripts.Script):
         if getattr(p, 'ip_adapter_masks', []) == []:
             p.ip_adapter_masks = args[MAX_ADAPTERS*5:MAX_ADAPTERS*6][:units]
             p.ip_adapter_masks = [x for x in p.ip_adapter_masks if x]
+        layers_active, layers = args[MAX_ADAPTERS*6:MAX_ADAPTERS*7]
+        if layers_active and len(layers) > 0:
+            try:
+                layers = json.loads(layers)
+                p.ip_adapter_layers = layers
+            except Exception as e:
+                shared.log.error(f'IP adapter: failed to parse layer scales: {e}')
         # ipadapter.apply(shared.sd_model, p, p.ip_adapter_names, p.ip_adapter_scales, p.ip_adapter_starts, p.ip_adapter_ends, p.ip_adapter_images) # called directly from processing.process_images_inner